rename: 예외처리 패키지 구조화
#23
