--- Application/Reward/src/main/java/com/example/Reward/Receipt/Exception/GiveStockErrorException.java
@@ -1,7 +1,0 @@
-package com.example.Reward.Receipt.Exception;
-
-public class GiveStockErrorException extends ReceiptException {
-    public GiveStockErrorException() {
-        super(ReceiptErrorStatus.GIVE_STOCK_ERROR);
-    }
-}
