--- Application/Reward/src/main/java/com/example/Reward/Receipt/Exception/KISApiException.java
@@ -1,7 +1,0 @@
-package com.example.Reward.Receipt.Exception;
-
-public class KISApiException extends ReceiptException {
-    public KISApiException() {
-        super(ReceiptErrorStatus.KIS_API);
-    }
-}
