--- Application/Reward/src/main/java/com/example/Reward/Receipt/Exception/NoStockException.java
@@ -1,7 +1,0 @@
-package com.example.Reward.Receipt.Exception;
-
-public class NoStockException extends ReceiptException {
-    public NoStockException() {
-        super(ReceiptErrorStatus.NO_STOCK);
-    }
-}
