+++ Application/Reward/src/main/java/com/example/Reward/Receipt/Exception/ReceiptExceptions/NoStockException.java
@@ -1,0 +1,10 @@
+package com.example.Reward.Receipt.Exception.ReceiptExceptions;
+
+import com.example.Reward.Receipt.Exception.ReceiptErrorStatus;
+import com.example.Reward.Receipt.Exception.ReceiptException;
+
+public class NoStockException extends ReceiptException {
+    public NoStockException() {
+        super(ReceiptErrorStatus.NO_STOCK);
+    }
+}
