+++ Application/Reward/src/main/java/com/example/Reward/Receipt/Exception/ReceiptExceptions/OcrErrorException.java
@@ -1,4 +1,7 @@
-package com.example.Reward.Receipt.Exception;
+package com.example.Reward.Receipt.Exception.ReceiptExceptions;
+
+import com.example.Reward.Receipt.Exception.ReceiptErrorStatus;
+import com.example.Reward.Receipt.Exception.ReceiptException;
 
 public class OcrErrorException extends ReceiptException {
     private String url;
