+++ Application/Reward/src/main/java/com/example/Reward/Receipt/Exception/ReceiptExceptions/S3UploadFailedException.java
@@ -1,0 +1,11 @@
+package com.example.Reward.Receipt.Exception.ReceiptExceptions;
+
+import com.example.Reward.Receipt.Exception.ReceiptErrorStatus;
+import com.example.Reward.Receipt.Exception.ReceiptException;
+
+public class S3UploadFailedException extends ReceiptException {
+    public S3UploadFailedException() {
+        super(ReceiptErrorStatus.S3_UPLOAD_FAILED);
+    }
+
+}
