--- Application/Reward/src/main/java/com/example/Reward/Receipt/Exception/S3UploadFailedException.java
@@ -1,8 +1,0 @@
-package com.example.Reward.Receipt.Exception;
-
-public class S3UploadFailedException extends ReceiptException {
-    public S3UploadFailedException() {
-        super(ReceiptErrorStatus.S3_UPLOAD_FAILED);
-    }
-
-}
